refactor: extract pr_state scanner.rs + InboxMessage::new_system builder
Task 1: Extract scan_and_emit, apply_gh_poll, apply_gh_observations,
and build_event_message from pr_state/mod.rs (2427 LOC) into a new
scanner.rs module (~280 LOC). Reducer + types stay in mod.rs.

Task 2: Add InboxMessage::new_system() constructor with chaining
builder methods (with_correlation_id, with_reviewed_head,
with_delivery_mode). Replace ~20 hand-constructed InboxMessage sites
across 19 files, eliminating ~890 lines of boilerplate.


Agend-Agent: fixup-dev-2
Agend-Branch: refactor/cicd-medium-findings-2
Agend-Issued-At: 2026-05-23T17:19:32.913352+00:00
