+++ src/daemon/cron_tick.rs
@@ -115,34 +115,7 @@ pub fn check_schedules(home: &Path, registry: &AgentRegistry) {
             let _ = crate::inbox::enqueue_with_idle_hint(
                 home,
                 target,
-                crate::inbox::InboxMessage {
-                    schema_version: 0,
-                    id: None,
-                    read_at: None,
-                    thread_id: None,
-                    parent_id: None,
-                    task_id: None,
-                    force_meta: None,
-                    correlation_id: None,
-                    reviewed_head: None,
-                    from: "system:schedule".to_string(),
-                    text: message.to_string(),
-                    kind: Some("schedule".to_string()),
-                    timestamp: now_utc.to_rfc3339(),
-                    channel: None,
-                    delivery_mode: None,
-                    attachments: vec![],
-                    in_reply_to_msg_id: None,
-                    in_reply_to_excerpt: None,
-                    superseded_by: None,
-                    from_id: None,
-                    broadcast_context: None,
-                    sequencing: None,
-                    eta_minutes: None,
-                    reporting_cadence: None,
-                    worktree_binding_required: None,
-                    pr_number: None,
-                },
+                crate::inbox::InboxMessage::new_system("system:schedule", "schedule", message),
             );
             "ok_inbox"
         };
