+++ src/daemon/helper_staleness_watchdog.rs
@@ -119,34 +119,13 @@ fn emit_staleness_alert(home: &Path, helper_name: &str) {
          `agend-terminal doctor` reports the same state.)"
     );
     for recipient in RECIPIENTS {
-        let msg = crate::inbox::InboxMessage {
-            schema_version: 0,
-            id: None,
-            from: "system:helper_staleness_watchdog".to_string(),
-            text: text.clone(),
-            kind: Some("helper_staleness_watchdog".to_string()),
-            timestamp: chrono::Utc::now().to_rfc3339(),
-            channel: None,
-            read_at: None,
-            thread_id: None,
-            parent_id: None,
-            delivery_mode: Some("inbox_fallback".to_string()),
-            task_id: None,
-            force_meta: None,
-            correlation_id: Some(helper_name.to_string()),
-            reviewed_head: None,
-            attachments: Vec::new(),
-            in_reply_to_msg_id: None,
-            in_reply_to_excerpt: None,
-            superseded_by: None,
-            from_id: None,
-            broadcast_context: None,
-            sequencing: None,
-            eta_minutes: None,
-            reporting_cadence: None,
-            worktree_binding_required: None,
-            pr_number: None,
-        };
+        let msg = crate::inbox::InboxMessage::new_system(
+            "system:helper_staleness_watchdog",
+            "helper_staleness_watchdog",
+            text.clone(),
+        )
+        .with_delivery_mode("inbox_fallback")
+        .with_correlation_id(helper_name);
         if let Err(e) = crate::inbox::enqueue_with_idle_hint(home, recipient, msg) {
             tracing::warn!(
                 error = %e,
