+++ src/daemon/poll_reminder.rs
@@ -105,28 +105,8 @@ mod tests {
                     id: Some(format!("m-{agent}-{i}")),
                     from: "test".into(),
                     text: format!("msg {i}"),
-                    kind: None,
                     timestamp: chrono::Utc::now().to_rfc3339(),
-                    channel: None,
-                    delivery_mode: None,
-                    force_meta: None,
-                    correlation_id: None,
-                    reviewed_head: None,
-                    read_at: None,
-                    thread_id: None,
-                    parent_id: None,
-                    task_id: None,
-                    attachments: vec![],
-                    in_reply_to_msg_id: None,
-                    in_reply_to_excerpt: None,
-                    superseded_by: None,
-                    from_id: None,
-                    broadcast_context: None,
-                    sequencing: None,
-                    eta_minutes: None,
-                    reporting_cadence: None,
-                    worktree_binding_required: None,
-                    pr_number: None,
+                    ..Default::default()
                 },
             );
         }
