+++ src/schedules.rs
@@ -886,34 +886,11 @@ mod tests {
             let _ = crate::inbox::enqueue(
                 &home,
                 &sched.target,
-                crate::inbox::InboxMessage {
-                    schema_version: 0,
-                    id: None,
-                    from: "system:schedule".to_string(),
-                    text: sched.message.clone(),
-                    kind: Some("schedule_replay".to_string()),
-                    timestamp: chrono::Utc::now().to_rfc3339(),
-                    channel: None,
-                    delivery_mode: None,
-                    force_meta: None,
-                    correlation_id: None,
-                    reviewed_head: None,
-                    read_at: None,
-                    thread_id: None,
-                    parent_id: None,
-                    task_id: None,
-                    attachments: vec![],
-                    in_reply_to_msg_id: None,
-                    in_reply_to_excerpt: None,
-                    superseded_by: None,
-                    from_id: None,
-                    broadcast_context: None,
-                    sequencing: None,
-                    eta_minutes: None,
-                    reporting_cadence: None,
-                    worktree_binding_required: None,
-                    pr_number: None,
-                },
+                crate::inbox::InboxMessage::new_system(
+                    "system:schedule",
+                    "schedule_replay",
+                    sched.message.clone(),
+                ),
             );
         }
