+++ src/verify.rs
@@ -310,32 +310,10 @@ fn test_inbox(home: &Path) -> TestResult {
             home,
             test_name,
             inbox::InboxMessage {
-                schema_version: 0,
-                id: None,
-                read_at: None,
-                thread_id: None,
-                parent_id: None,
-                task_id: None,
-                force_meta: None,
-                correlation_id: None,
-                reviewed_head: None,
                 from: format!("test-{i}"),
                 text: format!("msg {i}"),
-                kind: None,
                 timestamp: "2024-01-01T00:00:00Z".into(),
-                channel: None,
-                delivery_mode: None,
-                attachments: vec![],
-                in_reply_to_msg_id: None,
-                in_reply_to_excerpt: None,
-                superseded_by: None,
-                from_id: None,
-                broadcast_context: None,
-                sequencing: None,
-                eta_minutes: None,
-                reporting_cadence: None,
-                worktree_binding_required: None,
-                pr_number: None,
+                ..Default::default()
             },
         );
     }
